refactor plugin actions
