+++ src/features/PluginsUI/Render/StandaloneType/Iframe.tsx
@@ -94,9 +94,9 @@ const IFrameRender = memo<IFrameRenderProps>(({ url, id, payload, width = 600, h
   });
 
   // when plugin update state, we should update it to the message pluginState key
-  const updatePluginState = useChatStore((s) => s.updatePluginState);
+  const optimisticUpdatePluginState = useChatStore((s) => s.optimisticUpdatePluginState);
   useOnPluginStateUpdate((key, value) => {
-    updatePluginState(id, { [key]: value });
+    optimisticUpdatePluginState(id, { [key]: value });
   });
 
   // when plugin wants to get plugin settings, send it to plugin
